Completed unit tests for Exercise 2
